[FEAT] #58-방 입장 시 예외 처리

See merge request s12-final/S12P31A604!43

+++ backend/src/main/java/com/ssafy/backend/domain/participant/repository/ParticipantRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 
 public interface ParticipantRepository extends JpaRepository<Participant, Long> {
 	List<Participant> findByRoom(Room room);
-	long countByRoom(Room room);
+	int countByRoom(Room room);
 	boolean existsBySession(SessionEntity session);
 }
